bug TDI-24869 fixed:Logon Studio:after finish "Loading components", click "cancle", still able to logon Studio
git-svn-id: http://talendforge.org/svn/tos/trunk@102753 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.core.repository/src/main/java/org/talend/core/repository/model/ProxyRepositoryFactory.java
@@ -1756,7 +1756,9 @@ public final class ProxyRepositoryFactory implements IProxyRepositoryFactory {
                 currentMonitor.beginTask(Messages.getString("ProxyRepositoryFactory.exec.migration.tasks"), 1); //$NON-NLS-1$
                 executeMigrations(project, false, currentMonitor);
                 TimeMeasure.step("logOnProject", "executeMigrations(afterLogonTasks)");
-
+                if (monitor != null && monitor.isCanceled()) {
+                    throw new OperationCanceledException(""); //$NON-NLS-1$
+                }
                 // clean workspace
                 coreService.deleteAllJobs(false);
 
@@ -1769,9 +1771,15 @@ public final class ProxyRepositoryFactory implements IProxyRepositoryFactory {
                 } catch (SystemException e1) {
                     //
                 }
+                if (monitor != null && monitor.isCanceled()) {
+                    throw new OperationCanceledException(""); //$NON-NLS-1$
+                }
                 // rules
                 if (PluginChecker.isRulesPluginLoaded()) {
                     coreService.syncAllRules();
+                }
+                if (monitor != null && monitor.isCanceled()) {
+                    throw new OperationCanceledException(""); //$NON-NLS-1$
                 }
                 TimeMeasure.step("logOnProject", "sync repository (routines/rules/beans)");
 
@@ -1779,15 +1787,21 @@ public final class ProxyRepositoryFactory implements IProxyRepositoryFactory {
                 currentMonitor.beginTask(Messages.getString("ProxyRepositoryFactory.synchronizeLibraries"), 1); //$NON-NLS-1$
                 coreService.syncLibraries(currentMonitor);
                 TimeMeasure.step("logOnProject", "sync components libraries");
-
+                if (monitor != null && monitor.isCanceled()) {
+                    throw new OperationCanceledException(""); //$NON-NLS-1$
+                }
                 // sap
                 if (PluginChecker.isSAPWizardPluginLoaded()) {
                     coreService.synchronizeSapLib();
                 }
-
+                if (monitor != null && monitor.isCanceled()) {
+                    throw new OperationCanceledException(""); //$NON-NLS-1$
+                }
                 coreService.resetUniservLibraries();
                 TimeMeasure.step("logOnProject", "sync specific libraries");
-
+                if (monitor != null && monitor.isCanceled()) {
+                    throw new OperationCanceledException(""); //$NON-NLS-1$
+                }
                 // remove the auto-build to enhance the build speed and application's use
                 IWorkspace workspace = ResourcesPlugin.getWorkspace();
                 IWorkspaceDescription description = workspace.getDescription();
@@ -1798,7 +1812,9 @@ public final class ProxyRepositoryFactory implements IProxyRepositoryFactory {
                     // do nothing
                 }
                 coreService.createStatsLogAndImplicitParamter(project);
-
+                if (monitor != null && monitor.isCanceled()) {
+                    throw new OperationCanceledException(""); //$NON-NLS-1$
+                }
                 coreService.synchronizeMapptingXML();
 
                 if (monitor != null && monitor.isCanceled()) {
